requirements.txt
fastapi
uvicorn[standard]
torch
torchvision
numpy
pillow
python-multipart